separated textfield from questionpage, done
